common function for squeezing bytes out of sponge state
