+++ src/sponge.rs
@@ -1,12 +1,13 @@
 use crate::keccak;
+use std::cmp;
 
 /// Given N -bytes message, this routine consumes it into Keccak\[c\] permutation state s.t.
-/// offset ( second parameter ) denotes how many bytes are already consumed into rate portion
+/// `offset` ( second parameter ) denotes how many bytes are already consumed into rate portion
 /// of the state.
 ///
 /// - c i.e. capacity can be either of 256 or 512 -bits.
 /// - Rate portion will have bitwidth of 1600 - c.
-/// - offset will live in 0 <= offset < RATE_BYTES.
+/// - `offset` will live in 0 <= offset < RATE_BYTES.
 pub fn absorb<const RATE_BYTES: usize, const RATE_WORDS: usize>(
     state: &mut [u64; 25],
     offset: &mut usize,
@@ -60,13 +61,13 @@ pub fn absorb<const RATE_BYTES: usize, const RATE_WORDS: usize>(
     }
 }
 
-/// Given that N -bytes are already consumed into Keccak\[c\] permutation state, this routine
+/// Given that N message bytes are already consumed into Keccak\[c\] permutation state, this routine
 /// finalizes sponge state and makes it ready for squeezing, by appending padding bytes to input
 /// message s.t. total absorbed message byte length becomes multiple of RATE_BYTES.
 ///
 /// - c i.e. capacity can be either of 256 or 512 -bits.
 /// - Rate portion will have bitwidth of 1600 - c.
-/// - offset will live in 0 <= offset < RATE_BYTES.
+/// - `offset` will live in 0 <= offset < RATE_BYTES.
 pub fn finalize<const RATE_BYTES: usize, const RATE_WORDS: usize, const D: u8>(
     state: &mut [u64; 25],
     offset: &mut usize,
@@ -82,4 +83,40 @@ pub fn finalize<const RATE_BYTES: usize, const RATE_WORDS: usize, const D: u8>(
 
     keccak::permute(state);
     *offset = 0;
+}
+
+/// Given that Keccak\[c\] permutation state is finalized, this routine can be invoked
+/// for squeezing N -bytes out of rate portion of the state.
+///
+/// - c i.e. capacity can be either of 256 or 512 -bits.
+/// - Rate portion will have bitwidth of 1600 - c.
+/// - `readable` denotes how many bytes can be squeezed without permutating the sponge state.
+/// - When `readable` becomes 0, state needs to be permutated again, after which RATE_BYTES can be squeezed.
+pub fn squeeze<const RATE_BYTES: usize, const RATE_WORDS: usize>(
+    state: &mut [u64; 25],
+    readable: &mut usize,
+    out: &mut [u8],
+) {
+    let olen = out.len();
+    let mut rate = [0u8; RATE_BYTES];
+    let mut off = 0;
+
+    while off < olen {
+        let read = cmp::min(*readable, olen - off);
+        let soff = RATE_BYTES - *readable;
+
+        for i in 0..RATE_WORDS {
+            rate[i * 8..(i + 1) * 8].copy_from_slice(&state[i].to_le_bytes());
+        }
+
+        out[off..(off + read)].copy_from_slice(&rate[soff..(soff + read)]);
+
+        *readable -= read;
+        off += read;
+
+        if *readable == 0 {
+            keccak::permute(state);
+            *readable = RATE_BYTES;
+        }
+    }
 }
